flask_app/requirements.txt
flask==3.1.0
python-dotenv==1.0.1
flask-session==0.8.0
openai==1.12.0
werkzeug==3.1.3
jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0 